тест

+++ controllers/cards.js
@@ -27,12 +27,18 @@ const createCard = (req, res) => {
 };
 
 const deleteCardById = (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.cardId)) {
+    return res.status(400).send({
+      message: 'Invalid card ID',
+    });
+  }
+
   Card.findByIdAndDelete(req.params.id)
     .then((card) => res.status(200).send({
       card,
       message: "Card delete by id",
     }))
-    .catch((err) => res.status(404).send({
+    .catch((err) => res.status(400).send({
       message: 'Internal server error',
       err: err.message,
       stack: err.stack,
@@ -41,7 +47,7 @@ const deleteCardById = (req, res) => {
 
 const likeCard = (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(req.params.cardId)) {
-    return res.send({
+    return res.status(404).send({
       message: 'Invalid card ID',
     });
   }
@@ -59,7 +65,7 @@ const likeCard = (req, res) => {
 
 const deleteLike = (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(req.params.cardId)) {
-    return res.send({
+    return res.status(404).send({
       message: 'Invalid card ID',
     });
   }
